Orb 4.7 Code Tidy Up
Removal of code
Comment update / removed

+++ DX10 Framework/Files/DX10/DX10/2D Objects/DXSprite.cpp
@@ -127,12 +127,12 @@ int DXSprite::GetSliceHeight()
 
 float DXSprite::GetWidth()
 {
-	return m_imageWidth;
+	return (float)m_imageWidth;
 }
 
 float DXSprite::GetHeight()
 {
-	return m_imageHeight;
+	return (float)m_imageHeight;
 }
 
 float DXSprite::GetImageWidth()
